teter

+++ submit/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: akostrik <akostrik@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 14:21:18 by akostrik          #+#    #+#             */
-/*   Updated: 2022/11/18 18:06:25 by akostrik         ###   ########.fr       */
+/*   Updated: 2022/11/18 19:23:25 by akostrik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,12 +20,12 @@
 // the initial dst-length + src-length (to make truncation detection simple)
 // Replacements for strncat (more consistent, and less error prone)
 
-//tester ft_strlcat      : 1.KO 2.KO 3.KO 4.KO 5.KO 6.OK 7.KO 8.KO 9.KO 10.OK 11.OK 12.OK 13.KO 14.KO 15.OK 16.OK 17.KO 
+// 1.KO 2.KO 3.KO 4.KO 5.OK 6.KO 7.KO 8.KO 9.KO 10.OK 11.OK 12.OK 13.KO 14.KO 15.OK 16.OK 17.KO 
 // 1)  ft_strlcat(dest, src, 0) == strlen(src) && !strcmp(dest, "B")
 // 2)  ft_strlcat(dest, src, 1) == 10 && !strcmp(dest, "B")
 // 3)  ft_strlcat(dest, src, 3) == 3 + strlen(src) && !strcmp(dest, "BBBB")
 // 4)  ft_strlcat(dest, src, 6) == 13 && !strcmp(dest, "BBBBA")
-// 5)  ft_strlcat(dest, src, -1) == 14 && !strcmp(dest, "CCCCCAAAAAAAAA")
+// 6)
 // 7)  ft_strlcat(dest, src, 1) == strlen(src) && !strcmp(dest, "")
 // 8)  ft_strlcat(dest, src, 5) == strlen(src) + 5 && !strcmp(dest, "1111111111")
 // 9)  ft_strlcat(dest, src, 5) == strlen(src) + 5 && !strcmp(dest, "1111111111")
@@ -37,7 +37,7 @@
 
 static size_t	min(size_t a, size_t b)
 {
-	if ((int)a < (int)b)
+	if (a < b)
 		return (a);
 	return (b);
 }
@@ -50,11 +50,11 @@ static size_t	calc_nb_chars_to_append(char *dst, const char *src, size_t n)
 	dst_initial_len = ft_strlen(dst);
 	src_length = ft_strlen(src);
 
-	if ((int)dst_initial_len + (int)src_length + 1 <= (int)n)
+	if (dst_initial_len + src_length + 1 <= n)
 		return (dst_initial_len + src_length);
-	if ((int)dst_initial_len + 1 >= (int)n)
+	if (dst_initial_len + 1 >= n)
 		return (0);
-	return ((int)n - (int)dst_initial_len - 1);
+	return (n - dst_initial_len - 1);
 }
 
 size_t	ft_strlcat(char *dst, const char *src, size_t n)
